requirements.txt
google-cloud-speech>=2.27.0
google-cloud-storage>=2.18.0
moviepy>=1.0.3